refactor: move implementation to a folder
