+++ src/components/KeyboardAvoidingView/index.tsx
@@ -12,7 +12,7 @@ import Reanimated, {
   useDerivedValue,
   interpolate,
 } from 'react-native-reanimated';
-import { useKeyboardHandler } from '../hooks';
+import { useKeyboardAnimation } from './hooks';
 
 type Props = {
   /**
@@ -37,33 +37,6 @@ type Props = {
    */
   keyboardVerticalOffset?: number;
 } & ViewProps;
-
-const useKeyboardAnimation = () => {
-  const heightWhenOpened = useSharedValue(0);
-  const height = useSharedValue(0);
-  const progress = useSharedValue(0);
-
-  useKeyboardHandler(
-    {
-      onStart: (e) => {
-        'worklet';
-
-        if (e.height > 0) {
-          heightWhenOpened.value = e.height;
-        }
-      },
-      onMove: (e) => {
-        'worklet';
-
-        progress.value = e.progress;
-        height.value = e.height;
-      },
-    },
-    []
-  );
-
-  return { height, progress, heightWhenOpened };
-};
 
 const defaultLayout: LayoutRectangle = {
   x: 0,
